worker/pyproject.toml
[project]
name = "vlaarena-worker"
version = "0.1.0"
description = "Data aggregation worker for vlaarena"
requires-python = ">=3.11"
dependencies = [
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "apscheduler>=3.11.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "vlaarena-shared",
    "sqlalchemy>=2.0.44",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.2",
    "isort>=5.13.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["vlaarena_worker", "vlaarena_shared"]
force-single-line = false
lines-after-imports = 2
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
